stash: add t/t3908 demonstrating a failure with submodules
The new test t/t3908 demonstrates a case where 'git stash'
will irreversibly lose changes to modified submodule files
in a project with a submodule.

When the configuration option recurse.submodule is set to true,
'git stash' internally calls 'git reset --hard'
with the '--recurse-submodules' flag (implicitly).
If there are modified files in the submodule,
the changes will be lost as the submodule is reset to the commit
recorded in the superproject.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>