t2013: modernize submodule setup
The test 2013.1 ('setup') sets up a submodule in the trash directory
by creating a repository and adding it with `git add`.

This is not in line with current practice which is to use
`git submodule add` to add a submodule.

Let's use this command in the setup phase of the test to bring it
up to date. Let's also remove the subshells by using `test_create_repo`
and `test_commit -C`.

Using `git submodule add` has the effect of properly setting up the
submodule configuration in .gitmodules, so the `git config` calls
to set up submodule.submodule.path in tests 2013.5 and 2013.6 are
now unnecessary. Remove them.

Since .gitmodules is now tracked, the expected output in those two tests
needs to be changed as .gitmodules is modified by those tests and so it
appears in the output of `git checkout`.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>