lib-submodule-update: show tested command in test names again
In 5b0ac09fb1 (lib-submodule-update: pass 'test_must_fail' as an
argument, 2020-06-24), lib-submodule-update.sh was changed so that
'test_must_fail' is now only used with 'git' commands.

An unfortunate side-effect of that change, noted in the commit message,
is that 'git_test_func' now appears in the name of all tests since it is
the value of the 'command' variable that is used as a prefix for tests
names. Prior to that change 'command' contained the git function being
tested (or a test helper function) and so each test was prefixed with
the name of the command being tested.

Now that all tests are now prefixed with 'git_test_func', it is harder
to distinguish tests ran using 'test_submodule_switch',
'test_submodule_forced_switch' and 'test_submodule_switch_func' in the
output of test scripts.

Since 5b0ac09fb1 also introduced the global variable 'gitcmd' to hold
the real Git command (or test helper function) being tested, let's use
this as prefix in the test names, restoring the old output. We also need
to set it in 'test_submodule_switch_func', since it was not being
used in that function.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>