unpack-trees.h: add calling porcelain to 'struct unpack_trees_options'
In a subsequent patch we teach 'git read-tree' to accept a '--porcelain'
argument, so that commands that call 'git read-tree' can choose which
porcelain error messages to display.

Add a 'porcelain' string pointer to 'struct unpack_trees_options' and
set it in setup_unpack_trees_porcelain(), which already receives the
name of the porcelain command as a string.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>