unpack-trees: correctly super-prefix paths in porcelain messages
In an precedent commit we added the '--porcelain' flag to 'git
read-tree' and added it to the invocation in
'submodule.c::submodule_move_head' so that errors messages for paths in
the submodule now use the porcelain messages corresponding to the
command being run in the superproject.

During the 'git read-tree' invocation in the submodule(s), the functions
'display_error_msgs' and 'display_warning_msgs' call 'super_prefixed' on
the string 'path.buf', but at that point the rejected paths were already
prefixed with a tabulation, so the super-prefix is added before the
tabulation instead of before the path.

Fix this by calling 'super_prefixed' earlier on, in 'add_rejected_path',
when the rejected paths are added to 'unpack_rejects'.

WIP: ADD OUTPUT ?

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>