unpack-trees: drop 'Aborting' trailer when '--super-project' is active
In an precedent commit we added the '--porcelain' flag to 'git
read-tree' and added it to the invocation in
'submodule.c::submodule_move_head' so that errors messages for paths in
the submodule now use the porcelain messages corresponding to the
command being run in the superproject.

The output of the 'git read-tree' invocation in the submodule still ends
with 'Aborting', though, which is confusing to the user because it is
followed by the 'Could not update submodule' message and a second
'Aborting' line, when the git process in the super-project aborts.

To streamline the error message a little bit more, drop the 'Aborting'
trailer when unpack_trees is called from a command that was given
'--super-prefix'.

WIP: add output.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>