builtin/rebase: honor submodule.recurse
This works because it sets a global variable in submodule.c which is
read by unpack_trees.

WIP: we might want to be more careful and only activate it for the merge
backend...

WIP: we might need to add support for --recurse-submodules in 'git
stash' first, to support stashing submodules changes. This way if a
submodule is "modified (new commits)" before starting the rebase, we can
return it back to this state at the end of the rebase. We might want to
limit this to 'new commits' and not 'modified content' and 'untracked
content' in a first iteration, tough, for simplicity.