submodule-config: improve function comments
Make it clear that the object ID given to 'submodule_from_{path,name}',
'config_from' and 'config_from_gitmodules' can be 'null_oid()', and that
in that case the submodule is looked up from the worktree, the index or
HEAD (this follows the logic in 'config_from_gitmodules').

For 'config_from', make sure we reword the comment such that "on-demand"
loading only applies when reading from a revision (it is assumed that
'repo_read_gitmodules' was already called to populate the submodule
cache from the worktree, the index or HEAD).

Add a description for 'repo_read_gitmodules', which had none.

Finally, take the opportunity to align the arguments to
'submodule_from_{path,name}' in submodule-config.h.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>