submodule: add 'tree_submodule_from_ce'
The existing function 'submodule_from_ce' passes the null oid to
submodule_from_path(), such that only submodules present in the
worktree's .gitmodules are taken into account.

In a following commit we'll need a way to check the same conditions that
'submodule_from_ce()' verifies, but for a given tree instead.

Add 'tree_submodule_from_ce' that implements this and dispatch to it in
'submodule_from_ce'.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>