unpack-trees: add ERROR_MISSING_SUBMODULE
In a following commit, we want unpack_trees() to abort cleanly if a
submodule is not yet cloned but is configured to be recursively checked
out, to avoid aborting the checkout operation midway in check_updates().

Add a new plumbing error type, ERROR_MISSING_SUBMODULE, for this
situation. This is preferred to piggy-backing onto the existing
ERROR_WOULD_LOSE_SUBMODULE since that error refers to different
situation, namely a submodule already present in the worktree whose
local modifications would be lost.

In setup_unpack_trees_porcelain(), make use of the new
'submoduleNotUpdated' advice (see 961b130d20 (branch: add
--recurse-submodules option for branch creation, 2022-01-28) ) to try to
guide the user.

Note that since 'git submodule update' runs 'git checkout', we advise to
run it with '-c submodule.recurse=no' to avoid similar errors in the
case of missing nested submodules.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>