unpack-trees: check for missing submodule repository in merged_entry
Currently, the unpack_trees machinery is unprepared for recursively
checking out a submodule if that submodule is not yet cloned.

This leads to a rather bad user experience in the following situation:
- submodule 'sub' is absent at HEAD
- 'sub' is present in commit 'deadbeef'
- 'sub' was never cloned in the superproject clone
- a user recursively checks out 'deadbeef', either explicitely with
  '--recurse-submodules', or implicitely via a matching pathspec in
  'submodule.recurse' (usually the match-all pathspec '.' written by
  'git clone --recurse-submodules')

This fails when 'submodule_move_head' is called in 'write_entry' to
checkout 'sub', in the 'check_updates' phase of 'unpack_trees', with the
following message:

    fatal: not a git repository: ../.git/modules/sub
    fatal: could not reset submodule index

and leaves the worktree and repository in an inconsistent state:

- .gitmodules is written to disk an in 'deadbeef', but HEAD is
  unchanged, so it is either modified or untracked depending on the
  presence of other submodules at HEAD
- The sub directory in the worktree is created and the sub/.git gitfile
  is written to point to .git/modules/sub, which is is created but is
  empty apart from a 'config' file with 'core.wortree' pointing to 'sub'

The reason it fails in check_updates, instead of in the traverse_trees
phase where we check if the checkout operation is safe, is the
following:

1. In merged_entry, we invoke check_submodule_move_head only if
submodule_from_ce returns a valid submodule pointer.
2. submodule_from_ce passes 'null_oid()' as the treeish_name argument to
submodule_from_path, so the worktree, the index and HEAD are checked for
the presence of sub. The submodule is absent from all three, so we do not
get a valid submodule pointer and skip calling
check_submodule_move_head.
3. In check_updates, we call load_gitmodules_file, so the .gitmdodules
file is written early to the working tree as in commit 'deadbeef'.
4. In write_entry, we call submodule_from_ce again, and this time we get
a valid submodule pointer since .gitmodules was just updated in the
worktree.

Fix this bug by adding an additional check in 'merged_entry' by calling
'tree_submodule_from_ce' (added in a previous commit) and passing it the
object ID of the destination tree, which is available in
o->meta.treeish. This allows us to get a valid submodule pointer,
construct the path the the submodule repository, and abort with
ERROR_MISSING_SUBMODULE, added in the previous commit, if it is absent.

Reported-by: "Antoine W. Campagna" <awe@frontmatec.com>
Reported-by: Erik Claij <klimclay@gmail.com>
Reported-by: Luke Diamand <luke@diamand.org>
Reported-by: Mike Hommey <mh@glandium.org>
Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>