WIP: submodule.c: only return active submodules in tree_submodule_from_ce
This is necessary for the fix in subsequent commits to still pass the
test suite when there are some uninitialized submodules.

Obviously, this message should be improved because this change requires
a careful audit of each call to submodule_from_ce to see if it is
meaningful to skip non-active submodules.

The change could also be done in submodule_from_ce directly (i.e. before
the introduction of tree_submodule_from_ce in the previous commit).

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>