WIP: fix 'git merge --abort' when submodule is not yet cloned
We are in oneway_merge, we go into verify_uptodate_1. There we get a valid
submodule pointer from submodule_from_ce, so we go into
check_submodule_move_head and then submodule_move_head.

Then we get to this code:

        if (flags & SUBMODULE_MOVE_HEAD_FORCE)
                /*
                 * Pass non NULL pointer to is_submodule_populated_gently
                 * to prevent die()-ing. We'll use connect_work_tree_and_git_dir
                 * to fixup the submodule in the force case later.
                 */
                error_code_ptr = &error_code;
        else
                error_code_ptr = NULL;

        if (old_head && !is_submodule_populated_gently(path, error_code_ptr))
                return 0;

The reason we don't die here (despite not being in the forced case) is
that resolve_gitdir_gently (called by is_submodule_populated_gently)
does not die if the gitfile is simply missing. So we return 0, the
submodule is marked for updating, and we then die in check_updates as
"usual" ("could not reset submodule index").

So let's add an additional check in check_submodule_move_head: check
that the submodule is actually cloned, and reject the path with
ERROR_MISSING_SUBMODULE if not.

Funnily, this does not abort the 'git merge --abort' operation, but does
output error messages for the submodule (one from
check_submodule_move_head and one from verify_uptodate_1). This is
because in oneway_merge, verify_uptodate returns -1, so we do not set
the CE_UPDATE flag for the submodule, so it is skipped in check_updates.

It feels a bit weird that we do not abort the operation if
verify_uptodate returns an error... In ad17312e11 (unpack-trees:
oneway_merge to update submodules, 2018-01-05) it says "check if it is
up to date and set the flag if not", but in the case of a submodule,
verify_uptodate does not really check if a submodule is up to date, it
checks if it is OK to move from HEAD to old (so it's confusing).
So this means: if it is OK to move from HEAD to old, update the
submodule, if it is not OK, do not touch it.

Note that oneway_merge is the only caller of verify_uptodate that does
not abort if it returns an error...

Note that this fails the test "removed submodule absorbs submodules .git
directory" in t1013, t2013 and t7112 since these use a "non-absorbed"
submodule, and this patch does not account for that. We would have toi
either add the check in submodule_move_head and add a new error code to
mean ERROR_MISSING_SUBMODULE, or duplicate the
is_submodule_populated_gently check in check_submodule_move_head.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>