object: add repository argument to clear_object_flags
In a following commit we'll adjust reset_revision_walk, which calls
clear_object_flags, to clear the flags from the repository on which the
revision walk was performed, instead of hardcoding the_repository.

As a first step, add a 'struct repository' pointer argument to
clear_object_flags, and rename it to repo_clear_object_flags.

Adjust all callers, passing the appropriate repository argument if there
is one locally, and the_repository if not. Remove the UNUSED annotation
in upload_pack_v2 since the repository argument is now used.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>