merge-ort: add optional 'source' arg to format_commit
In a following commit we'll use 'format_commit' in a new function that
tries to improve the user experience for rebasing branches with
submodule changes. We will need to pass an additional string to the
function, which corresponds to the %S placeholder in pretty.h.

Add a 'source' argument now, and adjust the format used in the call to
repo_format_commit_message to add that information between brackets, if
it is present. Adjust the existing callers in merge_submodule
accordingly.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>