revision: introduce 'struct revision_source' for revision_sources slab
The revision_sources slab in 'struct rev_info' is a slab of pointers to
strings, which hold the ref names used in the 'git log --source' output.

These strings are strdup'ed in revision.c::handle_commit and are leaked
because we do not free them in release_revisions. Since in
revision.c::process_parents we set the 'slot' pointer to the address of
the string corresponding to the commit's parent in the slab, the
string pointers in the slab are shared and thus can't be easily freed.

To avoid leaking these strings, introduce a 'struct revision_source'
which holds a pointer to string as well as a 'count' integer, along with
appropriate REVISION_SOURCE_INIT macro and revision_source_init
function, and define the revision_sources slab as a slab of pointers to
'struct revision_source'.

Adjust revision.c::handle_commit and
builtin/fast-export.c::get_tags_and_duplicates to allocate a 'struct
revision_source', and update builtin/fast-export.c::handle_commit,
log-tree.c::show_decorations and pretty.c::format_commit_one, to access
the 'name' string of the 'struct revision_source' pointer.

Clear the revision_sources slab along with the other members of the
rev_info struct in release_revisions(), if it is present, by calling
'deep_clear_revision_source', passing in an appropriate
'free_revision_source' function which releases the strings when 'count'
reaches zero.

This avoids introducing a leak in several tests that are marked
leak-free when we add a revision walk using a revision_sources slab in a
following commit.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>