merge-ort.c: simplify logic in merge_submodule
In commit 24876ebf68 (commit-reach(repo_in_merge_bases_many): report
missing commits, 2024-02-28), the logic in merge_submodule was reworked
such that errors from repo_in_merge_bases would be reported. This was
done by cleverly using the same variable 'ret2' to hold the result of
both calls to repo_in_merge_bases, for the pairs (commit_o, commit_a)
and (commit_o, commit_b).

The existing logic however makes it harder to add aditionnal logic for
code that needs to be triggered if commit_a is behind commit_o, but
commit_b is ahead of commit_o, as will be done in the next commit.

Rework the logic by adding another variable ret3 to hold the return
value of the call to repo_in_merge_bases for the pair (commit_o,
commit_b). This gets rid of duplicated code for the error case. The
resulting logic is closer to what it was prior to 24876ebf68, except
that errors from repo_in_merge_bases are reported.

The only behaviour change is that instead of exiting immediately if the
first call to repo_in_merge_bases returns an error, we now exit
immediately after the second call.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>